Rename threatintel -> misp_stix_converter
Importing "threatintel" wasn't really clear in the sin you were
commiting.

Also, make some more functions so it ie easier to use from PyMISP
