+++ misp_stix_converter/converters/__init__.py
