+++ misp_stix_converter/converters/buildMISPAttribute.py
@@ -3,11 +3,10 @@ import re
 import cybox
 import logging
 import hashlib
-import json
 
 from pymisp import mispevent
 
-from threatintel.converters.lint_roller import lintRoll
+from misp_stix_converter.converters.lint_roller import lintRoll
 
 # Cybox cybox don't we all love cybox children
 from cybox.objects import email_message_object, file_object, address_object
@@ -38,7 +37,7 @@ def identifyHash(hsh):
     return possible_hashes
 
 
-def buildEvent(pkg, mispAPI, **kwargs):
+def buildEvent(pkg, **kwargs):
     log.info("Building Event...")
     if not pkg.stix_header:
         title = "STIX Import"
@@ -57,12 +56,7 @@ def buildEvent(pkg, mispAPI, **kwargs):
     for obj in lintRoll(pkg):
         # This will find literally every object ever.
         event = buildAttribute(obj, event)
-    if event.attributes:
-        response = mispAPI.add_event(json.dumps(event, cls=mispevent.EncodeUpdate))
-        if response.get('errors'):
-            # FIXME *maybe* we want to raise a thing there....
-            pass
-            # raise Exception(response.get('errors'))
+    return event
 
 
 def buildAttribute(pkg, mispEvent):
