+++ misp_stix_converter/converters/buildSTIXAttribute.py
