+++ misp_stix_converter/converters/convert.py
@@ -6,6 +6,7 @@
 # Sys imports
 import logging
 from tempfile import NamedTemporaryFile
+import json
 
 from pymisp import mispevent
 
@@ -15,9 +16,9 @@ from stix.core import STIXHeader
 from stix.indicator import Indicator
 
 # Local imports
-from threatintel.errors import STIXLoadError
-from threatintel.converters import buildSTIXAttribute
-from threatintel.converters import buildMISPAttribute
+from misp_stix_converter.errors import STIXLoadError
+from misp_stix_converter.converters import buildSTIXAttribute
+from misp_stix_converter.converters import buildMISPAttribute
 
 log = logging.getLogger("__main__")
 
@@ -30,8 +31,12 @@ def MISPtoSTIX(mispJSON):
         :returns stix: A STIX stix with as much of the original
                           data as we could convert.
     """
-    misp_event = mispevent.MISPEvent()
-    misp_event.load(mispJSON)
+    if isinstance(mispJSON, mispevent.MISPEvent):
+        misp_event = mispJSON
+    else:
+        misp_event = mispevent.MISPEvent()
+        misp_event.load(mispJSON)
+
     # We should now have a proper MISP JSON loaded.
 
     # Create a base stix
@@ -54,6 +59,29 @@ def MISPtoSTIX(mispJSON):
     return stix
 
 
+def load_stix(stix):
+    # Just save the pain and load it if the first character is a <
+
+    if isinstance(stix, STIXPackage):
+        return stix
+
+    f = NamedTemporaryFile(mode="w+")
+    f.write(stix)
+    f.seek(0)
+    # Oh no we have to try and load it now
+    try:
+        # Try loading from JSON
+        stix_package = STIXPackage().from_json(f.name)
+    except:
+        # Ok then try loading from XML
+        try:
+            stix_package = STIXPackage().from_xml(f.name)
+        except Exception as ex:
+            # No joy. Quit.
+            raise STIXLoadError("Could not load stix file. {}".format(ex))
+    return stix_package
+
+
 def STIXtoMISP(stix, mispAPI, **kwargs):
     """
         Function to convert from something stixxy ( as we have 3 possible representations )
@@ -63,23 +91,15 @@ def STIXtoMISP(stix, mispAPI, **kwargs):
     """
 
     log.info("Converting a package from STIX to MISP...")
-    # Just save the pain and load it if the first character is a <
-
-    if not isinstance(stix, STIXPackage):
-        f = NamedTemporaryFile(mode="w+")
-        f.write(stix)
-        f.seek(0)
-        # Oh no we have to try and load it now
-        try:
-            # Try loading from JSON
-            stix = STIXPackage().from_json(f.name)
-        except:
-            # Ok then try loading from XML
-            try:
-                stix = STIXPackage().from_xml(f.name)
-            except Exception as ex:
-                # No joy. Quit.
-                raise STIXLoadError("Could not load stix file. {}".format(ex))
-
+    stix = load_stix(stix)
     # Ok by now we should have a proper STIX object.
-    return buildMISPAttribute.buildEvent(stix, mispAPI)
+
+    misp_event = buildMISPAttribute.buildEvent(stix, **kwargs)
+    if misp_event.attributes:
+        response = mispAPI.add_event(json.dumps(misp_event, cls=mispevent.EncodeUpdate))
+        if response.get('errors'):
+            # FIXME *maybe* we want to raise a thing there....
+            pass
+            # raise Exception(response.get('errors'))
+
+        return response
