+++ misp_stix_converter/converters/lint_roller.py
