+++ misp_stix_converter/errors.py
