+++ setup.py
@@ -2,7 +2,7 @@
 
 # Setup script for ThreatIntel Conversion
 
-from setuptools import setup, find_packages
+from setuptools import setup
 import os
 
 setup(
@@ -12,12 +12,12 @@ setup(
     author="Hannah Ward",
     author_email="hannah.ward@baesystems.com",
     url="https://github.com/FloatingGhost/MISP-STIX-Converter",
-    packages=find_packages(),
+    packages=['misp_stix_converter'],
     install_requires=["pymisp>=2.4.50", "requests>=2.9.1", "pyaml>=3.11",
                       "stix>=1.2", "cybox>=2.0", "nose", "cabby"],
     dependency_links=["https://github.com/STIXProject/python-stix.git",
                       "https://github.com/CybOXProject/python-cybox.git"],
-    scripts=["threatintel/misp-to-stix.py", "threatintel/stix-to-misp.py"],
+    scripts=["misp_stix_converter/misp-to-stix.py", "misp_stix_converter/stix-to-misp.py"],
     data_files=[(os.path.expanduser("~/.misptostix"),
                 ['misp.login.example'])]
 )
