+++ tests/test_stix-to-misp.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 import glob
-from threatintel.converters import convert
-from threatintel.servers import misp
+from misp_stix_converter.converters import convert
+from misp_stix_converter.servers import misp
 
 
 def test_convert():
